requirements.txt
deep_translator==1.11.4
Django==5.0.4
djangorestframework==3.15.2
openai==1.52.0
python-dotenv==1.0.1
Requests==2.32.3
